Правки тестов

+++ migrations/000001_metrics.up.sql
@@ -3,4 +3,4 @@ CREATE TABLE metrics (
     name VARCHAR(255) NOT NULL,
     type VARCHAR(255) NOT NULL,
     value DOUBLE PRECISION,
-    delta INTEGER);
+    delta BIGINT);
